🎨 improve & optimize Factory

+++ src/Factory.sol
@@ -27,7 +27,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     ////////////////////////////////////////////////////////////// */
 
     // The latest Account version, newly deployed Account use the latest version by default.
-    uint16 public latestAccountVersion;
+    uint88 public latestAccountVersion;
     // The baseURI of the ERC721 tokens.
     string public baseURI;
 
@@ -50,7 +50,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
         address registry;
         // The contract address of the Account logic.
         address logic;
-        // Arbitrary data, can contain instructions to execute when updating Account to new logic.
+        // Arbitrary data, can contain instructions to execute when updating Account to new implementation.
         bytes data;
     }
 
@@ -58,9 +58,9 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
                                 EVENTS
     ////////////////////////////////////////////////////////////// */
 
-    event AccountUpgraded(address indexed accountAddress, uint16 indexed newVersion);
-    event AccountVersionAdded(uint16 indexed version, address indexed registry, address indexed logic);
-    event AccountVersionBlocked(uint16 version);
+    event AccountUpgraded(address indexed accountAddress, uint88 indexed newVersion);
+    event AccountVersionAdded(uint88 indexed version, address indexed registry, address indexed logic);
+    event AccountVersionBlocked(uint88 version);
 
     /* //////////////////////////////////////////////////////////////
                                 CONSTRUCTOR
@@ -81,7 +81,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
      * @return account The contract address of the proxy contract of the newly deployed Account.
      * @dev If accountVersion == 0, the newest version will be used.
      */
-    function createAccount(uint256 salt, uint16 accountVersion, address baseCurrency, address creditor)
+    function createAccount(uint256 salt, uint88 accountVersion, address baseCurrency, address creditor)
         external
         whenCreateNotPaused
         returns (address account)
@@ -135,7 +135,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
      * Checks are done such that only compatible versions can be upgraded to.
      * Merkle proofs and their leaves can be found on https://www.github.com/arcadia-finance.
      */
-    function upgradeAccountVersion(address account, uint16 version, bytes32[] calldata proofs) external {
+    function upgradeAccountVersion(address account, uint88 version, bytes32[] calldata proofs) external {
         if (_ownerOf[accountIndex[account]] != msg.sender) revert FactoryErrors.OnlyAccountOwner();
         if (accountVersionBlocked[version]) revert FactoryErrors.AccountVersionBlocked();
 
@@ -156,11 +156,12 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     }
 
     /**
-     * @notice Function used to transfer an Account between users.
+     * @notice Function used to transfer an Account between users based on Account address.
      * @param from The sender.
      * @param to The target.
      * @param account The address of the Account that is transferred.
-     * @dev This method transfers an Account not on id but on address and also transfers the Account proxy contract to the new owner.
+     * @dev This method transfers an Account on Account address instead of id and
+     * also transfers the Account proxy contract to the new owner.
      */
     function safeTransferFrom(address from, address to, address account) public {
         uint256 id = accountIndex[account];
@@ -169,11 +170,12 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     }
 
     /**
-     * @notice Function used to transfer an Account between users.
+     * @notice Function used to transfer an Account between users based on Account id.
      * @param from The sender.
      * @param to The target.
      * @param id The id of the Account that is about to be transferred.
-     * @dev This method overwrites the safeTransferFrom function in ERC721.sol to also transfer the Account proxy contract to the new owner.
+     * @dev This method overwrites the safeTransferFrom function in ERC721.sol to
+     * also transfer the Account proxy contract to the new owner.
      */
     function safeTransferFrom(address from, address to, uint256 id) public override {
         IAccount(allAccounts[id - 1]).transferOwnership(to);
@@ -181,12 +183,13 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     }
 
     /**
-     * @notice Function used to transfer an Account between users.
+     * @notice Function used to transfer an Account between users based on Account id.
      * @param from The sender.
      * @param to The target.
      * @param id The id of the Account that is about to be transferred.
      * @param data additional data, only used for onERC721Received.
-     * @dev This method overwrites the safeTransferFrom function in ERC721.sol to also transfer the Account proxy contract to the new owner.
+     * @dev This method overwrites the safeTransferFrom function in ERC721.sol to
+     * also transfer the Account proxy contract to the new owner.
      */
     function safeTransferFrom(address from, address to, uint256 id, bytes calldata data) public override {
         IAccount(allAccounts[id - 1]).transferOwnership(to);
@@ -194,11 +197,12 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     }
 
     /**
-     * @notice Function used to transfer an Account between users.
+     * @notice Function used to transfer an Account between users based on Account id.
      * @param from The sender.
      * @param to The target.
      * @param id The id of the Account that is about to be transferred.
-     * @dev This method overwrites the transferFrom function in ERC721.sol to also transfer the Account proxy contract to the new owner.
+     * @dev This method overwrites the transferFrom function in ERC721.sol to
+     * also transfer the Account proxy contract to the new owner.
      */
     function transferFrom(address from, address to, uint256 id) public override {
         IAccount(allAccounts[id - 1]).transferOwnership(to);
@@ -236,7 +240,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
         versionRoot = versionRoot_;
         versionInformation[latestAccountVersion_] = VersionInformation({ registry: registry, logic: logic, data: data });
 
-        emit AccountVersionAdded(uint16(latestAccountVersion_), registry, logic);
+        emit AccountVersionAdded(uint88(latestAccountVersion_), registry, logic);
     }
 
     /**
@@ -244,12 +248,14 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
      * @param version The Account version to be phased out.
      * @dev Should any Account logic version be phased out,
      * this function can be used to block it from being created for new Accounts.
+     * @dev Although possible to block an Account version through the versionRoot,
+     * that would require verifying the Merkle tree on account creation, which is expensive.
      */
     function blockAccountVersion(uint256 version) external onlyOwner {
         if (version == 0 || version > latestAccountVersion) revert FactoryErrors.InvalidAccountVersion();
         accountVersionBlocked[version] = true;
 
-        emit AccountVersionBlocked(uint16(version));
+        emit AccountVersionBlocked(uint88(version));
     }
 
     /*///////////////////////////////////////////////////////////////
@@ -271,7 +277,7 @@ contract Factory is IFactory, ERC721, FactoryGuardian {
     /**
      * @notice Function that stores a new base URI.
      * @param newBaseURI The new base URI to store.
-     * @dev tokenURI's of Arcadia Accounts are not meant to be immutable
+     * @dev tokenURIs of Arcadia Accounts are not meant to be immutable
      * and might be updated later to allow users to choose/create their own Account art,
      * as such no URI freeze is added.
      */
