+++ test/utils/Extensions.sol
@@ -131,7 +131,7 @@ contract ChainlinkOracleModuleExtension is ChainlinkOracleModule {
 contract FactoryGuardianExtension is FactoryGuardian {
     constructor() FactoryGuardian() { }
 
-    function setPauseTimestamp(uint256 pauseTimestamp_) public {
+    function setPauseTimestamp(uint96 pauseTimestamp_) public {
         pauseTimestamp = pauseTimestamp_;
     }
 
@@ -143,7 +143,7 @@ contract FactoryGuardianExtension is FactoryGuardian {
 contract RegistryGuardianExtension is RegistryGuardian {
     constructor() RegistryGuardian() { }
 
-    function setPauseTimestamp(uint256 pauseTimestamp_) public {
+    function setPauseTimestamp(uint96 pauseTimestamp_) public {
         pauseTimestamp = pauseTimestamp_;
     }
